# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/brokers": {
            "get": {
                "description": "get list of Kafka brokers from the sarama client",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Kafka",
                    "Broker"
                ],
                "summary": "List brokers",
                "responses": {
                    "200": {
                        "description": "List of broker details",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/cmd.BrokerDetail"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request when the host and port cannot be parsed",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/groups": {
            "get": {
                "description": "Get a list of all consumer groups from the Kafka cluster.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Kafka",
                    "ConsumerGroup"
                ],
                "summary": "List Kafka consumer groups",
                "responses": {
                    "200": {
                        "description": "OK: Map of consumer group names to their respective protocol types.",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request: Error message if unable to list groups.",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/publish": {
            "post": {
                "description": "Publishes a message to the specified Kafka topic using Sarama SyncProducer.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Kafka"
                ],
                "summary": "Publish message to Kafka topic",
                "parameters": [
                    {
                        "description": "Publish Body",
                        "name": "pubBody",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/cmd.PubBody"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Message is stored in topic(partition/offset)",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "error\": \"Error description",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/topics": {
            "get": {
                "description": "get list of all Kafka topics from the cluster",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Kafka",
                    "Topic"
                ],
                "summary": "List Kafka topics",
                "responses": {
                    "200": {
                        "description": "List of Kafka topics",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "$ref": "#/definitions/cmd.TopicDetail"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request when listing topics fails",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "cmd.BrokerDetail": {
            "type": "object",
            "properties": {
                "host": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "port": {
                    "type": "integer"
                },
                "rack": {
                    "type": "string"
                }
            }
        },
        "cmd.PubBody": {
            "type": "object",
            "required": [
                "content",
                "topic"
            ],
            "properties": {
                "content": {
                    "type": "string"
                },
                "topic": {
                    "type": "string"
                }
            }
        },
        "cmd.TopicDetail": {
            "type": "object",
            "properties": {
                "configEntries": {
                    "description": "ConfigEntries contains the custom topic configurations to set.",
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "numPartitions": {
                    "description": "NumPartitions contains the number of partitions to create in the topic, or\n-1 if we are either specifying a manual partition assignment or using the\ndefault partitions.",
                    "type": "integer"
                },
                "replicaAssignment": {
                    "description": "ReplicaAssignment contains the manual partition assignment, or the empty\narray if we are using automatic assignment.",
                    "type": "object",
                    "additionalProperties": {
                        "type": "array",
                        "items": {
                            "type": "integer"
                        }
                    }
                },
                "replicationFactor": {
                    "description": "ReplicationFactor contains the number of replicas to create for each\npartition in the topic, or -1 if we are either specifying a manual\npartition assignment or using the default replication factor.",
                    "type": "integer"
                }
            }
        }
    }
}